+++ mobile/android/android-components/docs/changelog.md
@@ -45,12 +45,13 @@ permalink: /changelog/
   * Add `PrivateNotificationFeature` to display a notification when private sessions are open.
 
 * **service-glean**
-  * Glean was updated to v31.2.2
+  * Glean was updated to v31.2.3
     * BUGFIX: Correctly format the date and time in the Date header
     * Feature: Add rate limiting capabilities to the upload manager
     * BUGFIX: baseline pings with reason "dirty startup" are no longer sent if Glean did not full initialize in the previous run
     * BUGFIX: Compile dependencies with `NDEBUG` to avoid linking unavailable symbols.
       This fixes a crash due to a missing `stderr` symbol on older Android.
+    * BUGFIX: Fix mismatch in events keys and values by using glean_parser version 1.23.0.
 
 * **feature-webnotifications**
   * `WebNotificationFeature` checks the site permissions first before showing a notification.